BloodPressure(IPSUR)	R Documentation

Blood Pressure and Heart Rate Readings

Description

These data were collected during from 2004 through 2006 by Taoying Bian.

Usage

data(BloodPressure)

Format

A data frame with 202 observations on the following 7 variables:

year

year. From 2004 through 2006

month

month of the year. January = 1.

day

the day of the month.

hour

the 24-clock hour.

systolic

systolic blood pressure reading (in mm Hg).

diastolic

diastolic blood pressure reading (in mm Hg).

heart.rate

heart rate reading, in beats per minute.

Details

From 2004 through 2006, Mrs. Taoying Bian regularly collected data concerning her blood pressure and heart rate.

Source

These data were collected by Taoying Bian from 2004 through 2006.

[Package IPSUR version 0.1-2 Index]