Confint(Rcmdr)	R Documentation

Confidence Intervals for Model Coefficients

Description

Except for glm objects, where a method is provided that provides intervals optionally based on the Wald statistic, this generic function simply calls confint in the stats package via its default method.

Usage


Confint(object, parm, level = 0.95, ...)

## S3 method for class 'glm':
Confint(object, parm, level=0.95, type=c("LR", "Wald"), ...)

Arguments

object 	 a model object.

parm 	 which parameters to use, defaults to all.

level 	 level of confidence, defaulting to 0.95.

type 	 for a glm object, confidence interval based on the profile likelihood (the default) or the Wald statistic.

... 	 arguments to be passed down to methods.

Value

dependent upon the method called.

Author(s)

John Fox jfox@mcmaster.ca

See Also

confint

[Package Rcmdr version 1.2-6 Index]