FeedingTimes(IPSUR)	R Documentation

Feeding Times of a Newborn

Description

These data were collected during July and August, 2006 at the request of the pediatrician concerning the feeding habits of Anna Lu Kerns.

Usage

data(FeedingTimes)

Format

A data frame with 42 observations on the following 7 variables:

age.days

age in days. July 1, 2006 = 1.

clock.hours

the 24-clock hour.

clock.min

the clock minute.

type

type of food eaten, being direct breast milk, formula, pumped breast milk, or no food (rest)

amount.oz

amount of food eaten, in ounces.

duration.min

duration of feeding time.

time.hours

sequential time in hours. Time = 0 corresponds to 8 AM, July 9th, 2006.

Details

During July and August 2006 the author collected data concerning the feeding habits of his newborn daughter, Anna Lu Kerns. The time of feeding was recorded, along with the type of food eaten. The amount of food eaten (in oz.) was recorded except when Anna was breastfeeding, in which case the duration of feeding was recorded. Some other durations were missing and others were calculated from the clock times.

Source

These data were collected by the author during July and August 2006 during observation of his newborn daughter.

[Package IPSUR version 0.1-2 Index]