Hist(Rcmdr)	R Documentation

Plot a Histogram

Description

This function is a wrapper for the hist function in the base package, permitting percentage scaling of the vertical axis in addition to frequency and density scaling.

Usage


Hist(x, scale = c("frequency", "percent", "density"), ...)

Arguments

x 	 a vector of values for which a histogram is to be plotted.

scale 	 the scaling of the vertical axis: "frequency" (the default), "percent", or "density".

... 	 arguments to be passed to hist.

Value

This function returns NULL, and is called for its side effect — plotting a histogram.

Author(s)

John Fox jfox@mcmaster.ca

See Also

hist

Examples


    data(Prestige)
    Hist(Prestige$income, scale="percent")
    

[Package Rcmdr version 1.2-6 Index]