IPSUR.Utilities(IPSUR)	R Documentation

IPSUR Utility Functions

Description

These functions support writing additions to the IPSUR package, and were patterned after their Rcmdr equivalents. Additional R code can be placed in files with file type .R in the etc subdirectory of the package. Add menus, submenus, and menu items by editing the file Rcmdr-menus.txt in the same directory.

Usage


checkMultiLevelFactors(n=1)
listMultiLevelFactors(dataSet=ActiveDataSet())
MultiLevelFactors(names)
multiLevelFactorsP(n=1)

Arguments

dataSet 	 the quoted name of a data frame in memory.

names 	 optional names to be stored.

n 	 number of variables to check for.

Details

There are several groups of functions exported by the Rcmdr package and documented briefly here. To see how these functions work, it is simplest to examine the dialog-generating functions in the Rcmdr package.

Checking for errors: The function checkMultiLevelFactors checks for the existence of objects and writes an error message to the log if it is absent (or insufficiently numerous, in the case of different kinds of variables).

Information: The following function returns vectors of object names: listMultiLevelFactors

Author(s)

G. Jay Kerns gkerns@ysu.edu

References

T. Lumley (2001) Programmer's niche: Macros in R. R News, 1(3), 11–13.

[Package IPSUR version 0.1-2 Index]