IPSURgetAnswer(IPSUR)	R Documentation

Display answers to selected problems

Description

Displays answers to selected problems in the system's web browser. Adapted from a function in the UsingR package by John Verzani.

Usage


IPSURgetAnswer(chapter = NULL, problem = NULL)

Arguments

chapter 	 The chapter number

problem 	 The problems number

Not all answers are available.

Details

Some selected answers from the problems in Introduction to Probability & Statistics Using R are available from the webpage http://www.cc.ysu.edu/~gjkerns/IPSUR. This function will display them one-by-one in the browser.

Value

If available, opens web browser to the requested answer.

Author(s)

G. Jay Kerns

See Also

See AlsoIPSURweb

Examples


IPSURgetAnswer()

[Package IPSUR version 0.1-2 Index]