IPSURweb(IPSUR)	R Documentation

Opens browser to IPSUR webpages

Description

Opens the browser to webpages from the IPSUR website. Adapted from a function in the UsingR package by John Verzani.

Usage


IPSURweb(what = c("homepage", "errata", "changes", "exercises", "package"))

Arguments

what 	 A character string indicating what page to open. No value specified will open the IPSUR homepage. Others are "errata" for typos and errors, "changes" for changes to the text brought about by new changes to R, "exercises" to go to a page containing selected answers, and "packages" for materials related to the IPSUR package.

Value

Opens the browser to the respective web page.

Author(s)

G. Jay Kerns

References

see http://www.cc.ysu.edu/~gjkerns

See Also

See AlsoIPSURgetAnswer

Examples


## Not run: IPSURweb()                    # main webpage
## Not run: IPSURweb("err")                       # errata

[Package IPSUR version 0.1-2 Index]