Rcmdr-package(Rcmdr)	R Documentation

R Commander

Description

A platform-independent basic-statistics GUI (graphical user interface) for R, based on the tcltk package.

Details

Package: 	 Rcmdr

Version: 	 1.2-6

Date: 	 2006/12/20

Depends: 	 R (>= 2.1.0), tcltk, grDevices, utils

Suggests: 	 abind, car (>= 1.1-1), effects (>= 1.0-7), foreign, grid, lattice, lmtest, MASS, mgcv, multcomp, nlme, nnet, relimp, rgl, RODBC

LazyLoad: 	 no

License: 	 GPL version 2 or newer

URL: 	 http://www.r-project.org, http://socserv.socsci.mcmaster.ca/jfox/Misc/Rcmdr/

Index:

Commander               R Commander
Compute                 Rcmdr Compute Dialog
Confint                 Confidence Intervals for Model Coefficients
Hist                    Plot a Histogram
KMeans                  K-Means Clustering Using Multiple Random Seeds
Rcmdr.Utilities         Rcmdr Utility Functions
Rcmdr.sciviews-specific
                        Rcmdr SciViews-specific Functions
RcmdrPager              Pager for Text Files
Recode                  Rcmdr Recode Dialog
Scatter3DDialog         Rcmdr 3D Scatterplot Dialog
aboutRcmdr              About the Rcmdr Package
assignCluster           Append a Cluster Membership Variable to a
                        Dataframe
bin.var                 Bin a Numeric Varisible
colPercents             Row, Column, and Total Percentage Tables
generalizedLinearModel
                        Rcmdr Generalized Linear Model Dialog
hierarchicalCluster     Rcmdr Hierarchical Clustering Dialog
linearModel             Rcmdr Linear Model Dialog
partial.cor             Partial Correlations
plotMeans               Plot Means for One or Two-Way Layout
reliability             Reliability of a Composite Scale
scatter3d               Three-Dimensional Scatterplots and Point
                        Identification
stem.leaf               Stem-and-Leaf Display

Translations

The R Commander comes with translations from English into several other languages. I am grateful to the following individuals for preparing these translations: Catalan, Manel Salamero; Spanish, Carlos Enrique Carleos Artime; French, Philippe Grosjean; Italian, Stefano Calza; Japanese, Takaharu Araki; Brazilian Portuguese, Adriano Azevedo Filho; Romanian, Adrian Dusa; Russian, Alexey Shipunov; Slovenian, Jaro Lajovic.

Author(s)

John Fox <jfox@mcmaster.ca>, with contributions from Michael Ash, Theophilius Boye, Stefano Calza, Andy Chang, Philippe Grosjean, Richard Heiberger, G. Jay Kerns, Renaud Lancelot, Matthieu Lesnoff, Martin Maechler, Dan Putler, Miroslav Ristic, and Peter Wolf.

Maintainer: John Fox <jfox@mcmaster.ca>

[Package Rcmdr version 1.2-6 Index]