RcmdrPager(Rcmdr)	R Documentation

Pager for Text Files

Description

This is a slightly modified version of the tkpager, changed to use the Rcmdr monospaced font and a white background.

Usage


RcmdrPager(file, header, title, delete.file)

Arguments

file 	 character vector of file(s) to be displayed.

header 	 for the beginning of each file.

title 	 for window

delete.file 	 delete file(s) on close.

See Also

tkpager

[Package Rcmdr version 1.2-6 Index]