RcmdrTestDrive(IPSUR)	R Documentation

Test-driving the R Commander

Description

These are simulated data specifically designed to allow the inexperienced user to browse the capabilities of the R Commander.

Usage

data(RcmdrTestDrive)

Format

A data frame with 168 observations on the following 9 variables:

Order

sequential order

Smoking

smoking status

Gender

gender of victim

Race

race of victim

Before

life expectancy before exposure

After

life expectancy after exposure

Salary

salary at retirement

Reduction

potential salary reduction

Parking

number of unpaid parking tickets

Details

The R Commander has extensive functionality, but many options are unavailable unless the correct types of data are loaded in the Active Data Set. This data set was randomly generated so that, when loaded, essentially all R Commander options would be available for the student to investigate. These data are entirely fictional. For an amusing contributed story tying these variables together, please visit http://www.cc.ysu.edu/~gjkerns/IPSUR/package.

Source

These data were randomly generated using the IPSUR probability menu for the R Commander.

[Package IPSUR version 0.1-2 Index]