colPercents(Rcmdr)	R Documentation

Row Column and Total Percentage Tables

Description

Percentage a matrix or higher-dimensional array of frequency counts by rows, columns, or total frequency.

Usage


colPercents(tab, digits=1)
rowPercents(tab, digits=1)
totPercents(tab, digits=1)

Arguments

tab 	 a matrix or higher-dimensional array of frequency counts.

digits 	 number of places to the right of the decimal place for percentages.

Value

Returns an array of the same size and shape as tab percentaged by rows or columns, plus rows or columns of totals and counts, or by the table total.

Author(s)

John Fox jfox@mcmaster.ca

[Package Rcmdr version 1.2-6 Index]