generalizedLinearModel(Rcmdr)	R Documentation

Rcmdr Generalized Linear Model Dialog

Description

This dialog is used to specify a generalized linear model to be fit by the glm function.

Details

The left model-formula box specifies the response variable to be used in the model; it may be a variable name or an expression evaluating to the response variable, such as working == "Fulltime".

The right model-formula box specifies the right-hand (i.e., predictor) side of the model. See glm for details.

You can type directly in the model formula boxes. Alternatively, double-clicking the left mouse button on a variable in the variable-list transfers it to the left-hand side of the model (if it is empty) or to the right-hand side. Factors are indicated in the variable list; all other variables are numeric. You can also enter operators and parentheses using the buttons above the formula.

Double-click the left mouse button to select a family in the "Family" box and the corresponding permissible link functions appear in the "Link function" box to the right. Initially, the canonical link for the family is selected. See family for details.

Specifying a subset expression allows you to fit the model to a subset of observations in the active data set. For example, assuming that gender is a variable in the active data set, entering gender == "Male" would restrict the model to males.

If the active model is a generalized linear model, and the active data set has not changed, then the initial values of the left-hand-side, right-hand-side, family, link, and subset fields are retained from the active model.

Author(s)

John Fox jfox@mcmaster.ca

See Also

glm, family, Comparison

[Package Rcmdr version 1.2-6 Index]