numSummary(Rcmdr)	R Documentation

Mean Standard Deviation and Quantiles for Numeric Variables

Description

numSummary creates neatly formatted tables of means, standard deviations, and quantiles of numeric variables.

Usage


numSummary(data, statistics=c("mean", "sd", "quantiles"), 
    quantiles=c(0, .25, .5, .75, 1), groups)
    
## S3 method for class 'numSummary':
print(x, ...)

Arguments

data 	 a numeric vector, matrix, or data frame.

statistics 	 any of "mean", "sd", or "quantiles", defaulting to all three.

quantiles 	 quantiles to report; default is c(0, 0.25, 0.5, 0.75, 1).

groups 	 optional variable, typically a factor, to be used to partition the data.

x 	 object of class "numSummary" to print.

... 	 arguments to pass down from the print method.

Value

numSummary returns an object of class "numSummary" containing the table of statistics to be reported along with information on missing data, if there are any.

Author(s)

John Fox jfox@mcmaster.ca

See Also

mean, sd, quantile.

Examples


library(car)
Prestige[1, "income"] <- NA
numSummary(Prestige[,c("income", "education")])
numSummary(Prestige[,c("income", "education")], groups=Prestige$type)
remove(Prestige)

[Package Rcmdr version 1.2-6 Index]