plotMeans(Rcmdr)	R Documentation

Plot Means for One or Two-Way Layout

Description

Plots cell means for a numeric variable in each category of a factor or in each combination of categories of two factors, optionally along with error bars based on cell standard errors or standard deviations.

Usage


plotMeans(response, factor1, factor2,
    error.bars = c("se", "sd", "conf.int", "none"), level=0.95,
    xlab = deparse(substitute(factor1)),
    ylab = paste("mean of", deparse(substitute(response))), 
    legend.lab = deparse(substitute(factor2)), main = "Plot of Means", 
    pch = 1:n.levs.2, lty = 1:n.levs.2, col = palette())

Arguments

response 	 Numeric variable for which means are to be computed.

factor1 	 Factor defining horizontal axis of the plot.

factor2 	 If present, factor defining profiles of means

error.bars 	 If "se", the default, error bars around means give plus or minus one standard error of the mean; if "sd", error bars give plus or minus one standard deviation; if "conf.int", error bars give a confidence interval around each mean; if "none", error bars are suppressed.

level 	 level of confidence for confidence intervals; default is .95

xlab 	 Label for horizontal axis.

ylab 	 Label for vertical axis.

legend.lab 	 Label for legend.

main 	 Label for the graph.

pch 	 Plotting characters for profiles of means.

lty 	 Line types for profiles of means.

col 	 Colours for profiles of means

Value

The function invisibly returns NULL.

Author(s)

John Fox jfox@mcmaster.ca

See Also

interaction.plot

[Package Rcmdr version 1.2-6 Index]