reliability(Rcmdr)	R Documentation

Reliability of a Composite Scale

Description

Calculates Cronbach's alpha and standardized alpha (lower bounds on reliability) for a composite (summated-rating) scale. Standardized alpha is for the sum of the standardized items. In addition, the function calculates alpha and standardized alpha for the scale with each item deleted in turn, and computes the correlation between each item and the sum of the other items.

Usage


reliability(S)

## S3 method for class 'reliability':
print(x, digits=4, ...)

Arguments

S 	 the covariance matrix of the items; normally, there should be at least 3 items and certainly no fewer than 2.

x 	 reliability object to be printed.

digits 	 number of decimal places.

... 	 not used: for compatibility with the print generic."

Value

an object of class reliability, which normally would be printed.

Author(s)

John Fox jfox@mcmaster.ca

References

N. Cliff (1986) Psychological testing theory. Pp. 343–349 in S. Kotz and N. Johnson, eds., Encyclopedia of Statistical Sciences, Vol. 7. Wiley.

See Also

cov

Examples


data(DavisThin)
reliability(cov(DavisThin))

[Package Rcmdr version 1.2-6 Index]